requirements.txt
beautifulsoup4==4.13.3
fake_useragent==2.0.3
Pillow==11.1.0
pyTelegramBotAPI==4.26.0
Requests==2.32.3
